extension entity added

+++ Business/Concrete/FolderManager.cs
@@ -1,5 +1,6 @@
 ﻿using AutoMapper;
 using Business.Abstract;
+using Core.Utilities.Business;
 using Core.Utilities.Results;
 using DataAccess.Abstract;
 using Entities.Concrete;
@@ -23,13 +24,19 @@ public class FolderManager : IFolderService
         Folder folder = mapper.Map<Folder>(folderForCreateDto);
         folder.Size = 0;
         folder.CreateDateTime = DateTime.Now;
-        folder.Type = "folder";
         this.folderDal.Add(folder);
         return new SuccessResult();
     }
 
-    public IResult Delete(Folder folder)
+    public IResult Delete(FolderForDelete folderForDelete)
     {
+        var result = BusinessRules.Run(CheckIfChildFolderExist(folderForDelete.Id), 
+                          CheckIfFolderExist(folderForDelete.Id));
+        if (result != null)
+        {
+            return new ErrorResult(result.Message);
+        }
+        Folder folder = mapper.Map<Folder>(folderForDelete);
         this.folderDal.Delete(folder);
         return new SuccessResult();
     }
@@ -44,9 +51,9 @@ public class FolderManager : IFolderService
         return new SuccessDataResult<List<Folder>>(folderDal.GetList(f=>f.FolderId == folderId).ToList());
     }
 
-    public IDataResult<List<string>> GetFolderLocation(int folderId)
+    public IDataResult<List<string>> GetFolderLocationById(int folderId)
     {
-        return new SuccessDataResult<List<string>>(this.folderDal.GetFolderLocation(folderId));
+        return new SuccessDataResult<List<string>>(this.folderDal.GetFolderLocationById(folderId));
     }
 
     public IResult Update(Folder folder)
@@ -54,4 +61,25 @@ public class FolderManager : IFolderService
         this.folderDal.Update(folder);
         return new SuccessResult();
     }
+
+    public IResult CheckIfChildFolderExist(int folderId)
+    {
+        var result = GetChildFoldersByFolderId(folderId);
+        if (result.Data.Count != 0)
+        {
+            return new ErrorResult("Klasör Boşaltılmadan Silinemez.");
+        }
+        return new SuccessResult();
+    }
+
+    public IResult CheckIfFolderExist(int folderId)
+    {
+        var result = GetById(folderId);
+        if (result.Data == null)
+        {
+            return new ErrorResult("Böyle bir klasör bulunamadı.");
+        }
+        return new SuccessResult();
+    }
+
 }
